requirements.txt
asgiref==3.11.1
django==4.2.28
djangorestframework==3.16.1
python-dotenv==1.2.1
sqlparse==0.5.5
typing-extensions==4.15.0
tzdata==2025.3
